requirements.txt
# Python libraries needed for the project
paho-mqtt
python-dotenv
mysql-connector-python
